requirements.txt
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.0
cycler==0.11.0
Flask==2.0.2
fonttools==4.28.2
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
jieba==0.42.1
Jinja2==3.0.3
kiwisolver==1.3.2
lxml==4.6.4
MarkupSafe==2.0.1
matplotlib==3.5.0
numpy==1.21.4
outcome==1.1.0
packaging==21.3
Pillow==8.4.0
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
selenium==4.1.0
setuptools-scm==6.3.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
tomli==1.2.2
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
Werkzeug==2.0.2
wordcloud==1.8.1
wsproto==1.0.0
